Adds :file, a command to compile files all-at-once like :paste.
Lots of people get tripped up on this, and right now the only way to test code
that uses companion objects in the REPL is to use :paste, which is infuriating
when the data is already available in a file.

Arguably this should be a patch to modify how :load works, but I'm not sure how
many people depend on :load to work as "replay" versus the functionality I'm
adding here. I can imagine four courses of action:

  1. Merge this branch as-is
  2. Replace the existing :load with :file
  3. Same as #2 but rename :load to :replay, and :file to :load
  4. Reject this patch

I'm open to any of those except #4.